Mooved cleanup to a separate file
